Implement the lifetime validation
At least one of them. According to the docs. of Microsoft.IdentityModel.Tokens.TokenValidationParameters.LifetimeValidator, if the delegate value is not null, it will be used to validate token's lifetime REGARDLESS of the value of TokenValidationParameters.ValidateLifetime. In the next commit I'm going to try to isolate this to a function of its own, as the code is getting pretty complicated to read.