This prevents breaking current API
If TokenValidationParameters is null (this is, for all the client code that is using current public APIs) we default to the old validation - check the TimeSpan difference and fail if it is less than TimeSpan.Zero. However, if the client code is using one of the new TryDecode overloads with TokenValidationParameters arguments, the new mechanism taking into account clock skew kicks in!