Get rid of the InternalPayloadParser delegate
It was almost identical to public PayloadParser, and it was ALWAYS being called with the 'serializerOptions' member as a second parameter, so it could be easily merged with PayloadParser. By getting rid of it, I'll avoid a good bunch of new overloads now that I'm going to add overloads taking the TokenValidationParameters parameter.