Further flexibilice lifetime validation
A common practice during unit testing is to have a way of changing the temporal frame of the test in an easy manner. By providing client code a way of specifying current system time, unit tests can check things such as that their system works as expected when a token is used outside of its lifetime frame without much of a hacking.