This solves a security issue!
Checking the signature is more important than checking the lifetime (not before / expire). Imagine the following scenario: you validate a token and it is expired. You still need some content on its payload (maybe for testing purposes, maybe for some other thing) so you call JwtDecoder.GetPayloadJson (which doesn't check signature, it just decodes the payload part of the token). Because the signature was NOT verified at any part of the process (not in JwtDecoder.TryDecode nor in JwtDecoder.GetPayloadJson) you might be dealing with a tampered token... and don't know about it!
So I changed the order of the operations. First we verify the signature. If the signature is invalid, we return with the error. Then we verify the lifetime. And, if the token is outside its lifetime, we still have the parsed token at the 'out T payloadResult' out variable, so calling again to JwtDecoder.GetPayloadJson to get the payload in case the token is expired (but can be trusted because signature matches) is not necessary (and, at that point, we already did the work, so it is honestly not worth it to hide it from the caller...!)